+++ services/game_handler/include/game_handler/Hand.hpp
@@ -5,8 +5,8 @@
 namespace GameHandler {
     using enum Card::Rank;
 
-    static const int8_t PREMIUM_NUMBER    = 5;
-    static const int8_t HAND_CARDS_NUMBER = 2;
+    static const int32_t PREMIUM_NUMBER    = 5;
+    static const int32_t HAND_CARDS_NUMBER = 2;
 
     static constexpr std::array<std::array<Card::Rank, HAND_CARDS_NUMBER>, PREMIUM_NUMBER> PREMIUM = {
         {{{QUEEN, QUEEN}}, {{KING, ACE}}, {{ACE, KING}}, {{KING, KING}}, {{ACE, ACE}}}};
